Tools for working with udhr data.
The ohchr_all.html file was manually extracted from the ohchr.org site
html and copy/poasted into a form that extract_ohchr_attributions can
parse.  That tool turns it into a .tsv file that records our
understanding of which udhr text is official 'UN' data and which is
not.  update_udhr_samples extracts the actual data we use from the
unicode.org/udhr site.

Other than the ohchr_all.html file this is just tools, not the data
generated by the tools.  That data will be in a separate commit.