Tweak attributions generated by update_udhr_samples
In some cases, all the bcp47 tags with a given language and script
have region or variant subtags.  Since there's code that looks up the sample
and attribution data using only language and script, we need to provide
a corresponding bcp47 tag and map it to one of the existing alternatives.

This adds a table defining which existing alternative to use when generating
additional lang-script tags, and applies it.

In addition:
- clarify comments at top of attributions.txt
- sort output of attributions.txt
- add flag to prevent staging of files that were generated under the repo