Rename files to indicate they are based on udhr or arbitrary chars.
This set differs from the ones changed in the previous commit because these
files are not generated from tools.  These get a different log message that
will help differentiate files that can be overwritten by tools from those that
should not be.

In addition, this renames the file da-Latin.txt to correct the misspelling
of the script tag.