Change noto_lint to distinguish errors from warnings
- adds option '-nw' / '--nowarn' to suppress warnings.
- changes gdef/classdef test failures to be warnings instead of errors.
- changes CSV output to include a 'Type' column in position zero, with values
  of 'Error' or 'Warning' depending on the type of failure.
- changes plaintext output to prefix messages with 'E' or 'W' depending
  on the type of failure, and includes the test tag (used in config) in angle
  brackets before the main body of the message for ease of reference.
- reports the number of warnings in the per-file summary when non-zero and
  not suppressed.

In addition the code was refactored a bit so that test checks are done inside
the 'warn' function by default, since the warn function now gets the test
tag passed to it. Most tests are very lightweight (cheaper than testing
the tag, in fact) so performing the test before checking if a failure needs
to be reported makes sense, as well as being more parsimonious. Since there
are tests in which the check should be performed earlier (either before a
loop, or because particular glyphs/chars play a part in determining whether
to report the error), a check_tests parameter was added to the warn fn
so that this default behavior can be overridden.