Support Emoji in website data generation.
Because the Emoji font had been inadvertently omitted from fonts.conf,
generating a sample image of the Emoji font was failing.  This fixes
fonts.conf, updates the web data generation tool to include it in the
output, and updates the fallback 'exemplar' chars used for emoji to
include more (and only) characters supported by the font.

In addition, the web data generation tool was generating a lang count by
counting all langs with samples, rather than only those languages that
get listed.  (Emoji has no languages, but it was reporting 1).  The
inconsistency between this reported number and the actual number you see
when opening the sample view was a problem.  This was also fixed.