Tool to swat noto license (and other name table records).
This is a one-off to update the license Noto uses to OFL.  At the same
time we clean up a number of lint issues/small inconsistencies with a
few other name table fields.  It's not designed to be run on the CJK
fonts.  It needs a bit of tweaking still.