Change GSUB/GPOS whitelist to use script code instead of name.
Also cleans up script name code a bit.  We had been using the
unicode data based on PropertyValueAliases.txt for the script names.
As it turns out, this data is incomplete and doesn't cover all the
scripts.  There's no particular reason lint should restrict itself
to this set of scripts, so I also changed the code to fallback the
script name to the english script names from CLDR.  Since these
differ in a few cases from the names in PropertyValueAliases, I
still use those values as the primary source.