Enable noto_fonts to be used without notoconfig.
This used to expect notoconfig to have values for 'noto_fonts' etc.
Some other utilities (lint) use this file where it's not necessary to
have these directories, and they fail because of it.  Make these optional
and allow get_noto_fonts to be called with a user-supplied list of paths.