Move some data extraction methods from Roboto.
These methods extract out data from fonts' full names. Moved from
https://github.com/google/roboto/blob/master/scripts/roboto_data.py,
but made a bit more general and expanded to include methods for family
and style, for future tests.

The map of weights is still specific to Roboto, but could be easily
expanded with the weights from Noto (or any common weight name).