Split semantics of how script keys and 'meta' scripts are handled.
There used to be a distinction between a 'script key' which was a
code we made up to assign to fonts that supported multiple scripts,
and a regular script code.  In addition, the multiple scripts that
could be returned for a script key had a 'primary' script that was
the default script to use when viewing the font to which the key
was applied.

It turns out that this division of semantics isn't quite right.
Some fonts that use standard script codes, like 'Jpan', also support
multiple (unicode) scripts, for instance Kana and Hira, because
these script codes are themselves 'higher level' codes used to
describe writing systems and not characters.  Since we need to
know that a font supports languages written in Kana (e.g. Ainu)
we neede to extend this notion of mapping a single code to multiple
codes to include standard script codes and not simply our keys.
Yet it seems best to segregate support for real scripts from the
keys we're using internally for convenience.

In addition it seemed inappropriate to overload this function with
a separate function of defining a preferred script for fonts (or
keys) that map to multiple scripts.  So that too was split into
a separate function.