Describe Okinawan as ryu-Jpan instead of ryu-Kana.
The main import of this is to affect the sample used by the web
site to illustrate Okinawan.  We don't have an Okinawan text sample
so use collections of characters based on the script.  Some
traditionalists use hiragana + kanji to write Okinawan according to
Wikipedia, so we just show the broader set based on Jpan instead of
restricting the set to Kana.