Add a tool for generating text samples based on glyphs in a font.
Some bug reports refer to names of characters in a particular script, but
don't provide the unicode values of these characters.  This tool lets
you interactively enter (portions of) names of characters and match these
names against the unicode names of characters available in the font to
build up a text string, which you can then write to a file.  This makes
it easier to build example text to view in that font.