Keep exemplar data in original sort order.
This also fixes a bug in which we cached the sorted exemplar list, but
returned (when not fetching from cache) the unsorted exemplar set.

Exemplars are recommended to be in collation order, but are actually sets.
For some of this data we're not guaranteed to be able to reconstruct the
collation order, so rather than destroy it and try to recreate it, or
use unicode order, we just assume the exemplar order is what we want.

The exemplar data is used in some cases to provide text samples for the
web site where we don't have actual text, they're lists of characters.
This makes it more likely that the (language-specific) lists have the
expected order.