Add tools to generate and compare cmap data from various sources.
The cmap data is organized per script and written to an xml format.  It
is sourced from a number of places: (a copy of) the lint code, the noto font
families, planning spreadsheets from MTI.  There are separate tools to
extract data from each source, a common set of utilities to read/write
the xml files, and a tool to compare the files.