Update lint name check to use noto_names data.
This renames FamilyNameData to NameTableData and adds the other name
table fields that lint cares about to it.  Methods that computed the
expected values of these fields were moved to / reimplemented in
noto_names, as were some of the constants (the license text and urls,
copyright line, etc.).  The expectations were tightened somewhat to
match values resulting from the swat work that was done for the
copyright update.  The lint check_name_table fn was refactored to rely
on the data from noto_names and unify the formatting of error
messages.

This updates noto_fonts to add a 'manufacturer' field to NotoFont and
compute values for it. (Other fields were added in a different commit
to support the phase 3 names).

This adds a '--phase' option to noto_lint.  Currently it only
controls whether preferred and wws name fields are checked (not if phase
is 2 or less).  Defaults to phase 3.

Some of the lint config test names were changed so that they would read
more naturally when converted to running text in error messages for
the noto names check.  A few tests were removed as they were subsumed
in a different test, and new test names were added for new name fields.