Put 'mono' into its own script/family.
This is something of a hack.  Currently we collect cmaps by script, but
we want to break out the monospace font's cmap separately since it's not
necessarily the same as LGC's.  The cmap analysis code has to be changed
to incorporate more aspects of the font besides script-- ideally, this
would align with a clear definition of 'family'.  Not there yet.