Add ability to output csv to mti_cmap_data tool.
Because Google Spreadsheets still (!) does not provide an option to prevent
autoconversion of text to numbers, this puts 'U+' in front of each cell
so that Google Spreadsheets will leave it alone.  Then afterwards you can
mark all columns as text only in Google Spreadsheets, and replace all to remove
the leading U+.  Otherwise some of the hex is interpreted as decimal numbers,
losing the leading zeros, and some as text.  Urg.