Tool to generate noto phase 3 cmap requests.
This tool is a new take on the collection of data that's currently in
noto_data.py.  It was difficult, with noto_data, to know why any particular
character was there.  Furthermore, the actual requirements depended on
both the various pieces of data in this file, plus the code in noto_lint
and lint_cmap_reqs that used them.  The new tool bases the requirements
on a sequence of steps that add to and remove characters from various
scripts, optionally logging operations so you follow the process at an
individual character level.  The output is a cmap xml file.

The intent is to then check in the phase 3 file generated from this tool,
and a phase 2 file based on the previous phase 2 fonts, and then use
them when checking cmap requirements.  noto_data will be significantly
reduced or removed.

This checkin retains some code created during development but not currently
used, to get it into the source history.  It will be deleted in subsequent
commits.