Expand definition of excluded codepoints.
'EXCL' is for codepoints which are (in the main) excluded from all fonts.
This makes it more explicit that surrogate components, private use, and
variation selectors are excluded.

It also reassigns Tags to excluded, formerly they were Symbols2.

Skips warning about reassigning characters that are not assigned in the
case of EXCL, we don't assume these have been assigned at this point.