Command line tool with params to subset fonts using cmap.
subset.py does something similar, but provides no command-line options
in its main-- it's intended to be used from a makefile, or called from
other code.

This doesn't use subset.py since it also provides an option to bump the
version number, and that's more straightforward to do before writing the
file than by reading it back in and writing it out a second time.  subset.py
is itself just a wrapper around the fonttools subset functionality so
there isn't much duplication.