Don't include alpha by default when scanning noto fonts.
Formerly this used to just have historical and Nastaliq.  Nastaliq
is no longer there, so we don't need to scan alpha to get it, and
the preliminary versions of phase 3 fonts have been added, and we
don't want them to be accessed by default (e.g. by tools to generate
the noto website data).