Update generate_emoji_name_data to order and group emoji.
This uses emoji_ordering to determine the grouping and ordering of emoji,
whereas before it just output a single list sorted by codepoint.  The data
format is now a list of dicts with category name and the list of emoji
data for that category.

This also tries to generate a name for all the emoji sequences it outputs,
formerly it provided no names for several sequences.  Related variants
(couples, families, skin tone sequences) get the same generic name.

Finally, this adds aliases from the NotoColorEmoji font for regions that
share the same flag as another region.