Remove some unwanted characters from the indic fonts.
Formerly, when computing the cmaps a set of characters was added to all
the indic fonts.  This was based on an incomplete understanding of what
was in the existing fonts.  As it turns out, we already add code points
from existing fonts to the new required sets in a later step, and this
correctly captures the information the original step was attempting to
provide.  The fix is to just delete the earlier step and not add the
extra characters.