Update proposed emoji data.
Emoji proposed for Unicode 9 all have names in UnicodeData.txt, so
don't need special-casing.  The ones that are candidates for a later
release might still be useful, so the APIs remain.  This renames the
data file and replaces it with the "Post U9.0 Candidates" from the
upstream unicode document.  The data changed a bit since the previous
snapshot so the regex changed.