Ensure we can parse proposed-emoji-10 data.
The parsing code wasn't updated properly when the proposed data was
updated, causing parsing to fail when code asked for the new data.
This fixes that.  The proposed data (copied from website) has some
extra lines in it that conveniently all start with the same character,
so a check was added to treat this character like a comment.  One
line was removed since there was no easy syntatic identifier to allow
for skipping it.