Manually add Cherokee UDHR article 1 to sample_texts.
The UDHR data has this at stage 1 (no known complete translation) and not
at stage 3 (partial Unicode conversion, often only article 1) and the tooling
filters out source data at state 2 or earlier, so it skips this even though
there is (and only is) a complete Article 1 text.  Rather than update the
tooling it's simpler to just patch this.

This fixes #239.