[notodiff] Clean up reporting for shape diffing
Before there were options provided by the shape diff class for
returning a report or maintaining a separate passed-in report. Now
there's only the passed-in report (now passed in to init instead of
the diff methods).

This means we only have to sort diffs once, in the dump report
function (which has been moved to the shape diff class from the
notodiff module).