Add tools to generate and report coverage data.
The generate tool takes directories and/or files and opens the
files, creating a set of all the codepoints in their cmaps.  It can
also take a cmap_data file.  It writes the data to an xml file which
is basically just a name and the ranges representing the codepoint set,
plus some metadata.

The report tool takes one or more of these coverage.xml files and
generates a report of coverage by unicode block.  Reports can be
text, csv, or html.  Coverage can be against a particular version
of Unicode, defaulting to 6 (noto phase 2 target).