Implement CodeList abstraction over data used by char comparison generator.
Sometimes we deal with cmaps, sometimes with lists of codes, sometimes with
lists of either one or two codes.  Wrap the different representations up
in a utility class to provide a common API, and provide different
constructors for it.

This has different implementations depending on the data we have, might
not really need to bother.