Refactor font data in character comparison html generator.
Formerly data was split into reference (really comparison) fonts and
target fonts.  Each had slightly different sets of data.  Target fonts
had a list of associated codepoints, for each target font we generated
a table showing each codepoint in all the comparison fonts and that
target font.  The comparison font data let you group a few fonts into
a single column and just used the fonts cmap to determine if it could
display the font.

This is restructured to split data into font data and codepoint data.
All font data is now structured like the comparison font data, except
this data can take the codepoint files we formerly used for the
target fonts.  This lets any font be a comparison font, including
the dingbats fonts that have PUA characters.

The target data is now a named list of codepoints, it can use the same
data as the target fonts used before or other cmap data.  We build
tables for each target by selecting the comparison fonts that have any
of the glyphs in the target set.

This new structure allows targets to be any list of related characters,
not only those associated with a particular font.

The main feature loss at the moment is the ability to specify exactly
which comparison fonts to use and their order.  Formerly the target
font was the rightmost, this made it easy to identify.  Also, the other
targets were omitted, in the case where you didn't want to see comparisons
between them (looked at another way, that was a limitation, since you
couldn't see comparisons between them).  Specifying the targets is now
also a bit more verbose.  These issues will be addressed.