Add NamesList.txt unicode data and apis to read it.
NamesList.txt is used for generating the unicode book.  The format is
designed for that purpose and not really as a data format.  However, this
is the only way to access information about related characters (that are
not equivalent) and also other names that some characters are known by
(e.g. LINE FEED).  This provides apis to get these two pieces of information
about a character, but does not attempt to parse the entirety of the
information from the NamesList.txt file.