Get alternate names from NameAliases.txt instead.
This is the formal place to provide name aliases, not NamesList which
is only for the unicode book.  Convert the previous code to get aliases
from this file instead.

This file also labels the names according to type.  Names with type
'figment' are ignored, others (control, abbreviation, correction,
alternate) are present.  The returned data has changed to be a
list of tuples so that callers can filter by type.