Add more emoji text variant support.
This changes the optional argument to get_unicode_emoji_variants to
a string.  There are some additional emoji we want in text form beyond
those in the Aug 2016 proposal, this allows for 'proposed' to include the
proposed variants, and 'proposed_extra' to include these additional ones
on top of that.  The default behavior remains the same as before.