Minor cleanup of unicode_data.
The load_data api now loads all the data, it hadn't been updated when
additional data files were added during the past year.  It is used
to preload all data rather than just load it on demand.

The name api now reports the codepoint when it fails to find a name
for it, formerly it didn't tell you.

Alt_names now has a docsctring describing the format of the data it
returns.