Remove leading 'U+' from generated spreadsheet data.
Google Sheets for the longest time provided no mechanism when uploading
data to not automatically convert strings to dates and numbers.  This
meant that hex strings (the codepoints) was arbitrarily converted to integers
depending on whether the string contained a non-digit or not.  Prepending
'U+' fixed this, but the spreadsheets originally didn't have this so after
uploading the data you needed to edit it to remove them.

Sheets now lets you optionally stop this unwanted conversion, so we no
longer need to generate the prefix.