Get character flag data from file.
Flags get added the name field of a character if it is in (or not in)
a set associated with the flag.

This was hard-coded in the tool, now it's driven by a data file.
It means I don't have to edit the tool to change the flags, and can
more easily switch between sets.  But it's a lot more code...