Parse ordering data from html.
The other ordering sources are not maintained, and there were no other
sources that include the categories and subcategories, so this adds
the ability to scrape the emoji ordering and category information from
the table on the html page on the unicode site.

This also adds a flag to check the emoji ordering data against a
directory containing emoji_u*.png files.  Unfortunately currently
slight differences in naming behavior cause more mismatches than we'd
like.