Update emoji name data tool.
This tool generates a data file used by the website to preview the
emoji images.  This change updates the tool to handle new emoji in
the upcoming android release, in particular gendered professions.

All gender and skin color variants of an emoji are given the same
name in the data file-- we just let the image differences speak for
themselves.  The name data we scrape from unicode sources doesn't
include these sequences yet, so this update provides names for them.
We will probably keep using these names even when/if we have data for
the additional sequences.

This adds gender information on existing 'simple' emoji that, in the
current NotoColorEmoji at least, are represented as having a gender.
There is no unicode data to indicate what gender they have, and they
don't necessarily have one, so we use the gender these have in the
NotoColorEmoji font.  When generating the file for the website, we
don't need to display these images since they _ought to be_ identical
to the corresponding gendered (and skin-toned) version. So we filter
out these 'simple' emoji, which also lets us use the gender-neutral
names we provide for the sequences (e.g. 'Person with Turban').

Finally, the latest NotoColorEmoji adds single-parent family sequences
and the rainbow flag, so this adds names for those as well.