Order Mono with Sans/Serif LGC.
This impacts the default order in which fonts are listed on the website.
This makes the top three fonts be Sans, Serif, and Mono; formerly just
Sans and Serif got this treatment.  Other fonts order alphabetically
following these.

This also makes a slight change to how one of the flags to skip rebuilding
parts of the data is treated.