Update font comparison target syntax for sequences, support _math_.
This refactors the target code into Target subclasses that generate
text or html output, most of the previous html table and text code
was moved into these subclasses.

This was to support a 'sequence' target that presents a strings
of text in a single target font, in addition to the tables of
individual characters across a set of fonts.  Formerly this was
hardcoded to generate sequences for Symbols2 combining marks, this
makes the sequence data-driven so it could be used for Symbols as
well.

The directive syntax in the target spec parser was updated to allow a
directive that specifies the type of target, with a default of 'table'
to support the existing syntax.  The new sequence target takes slightly
different arguments.

The sequence target is still a bit hacky, it specifically calls out a
'suffix' to insert between codepoints in each string-- this is because
we use the strings mostly to illustrate combining marks and this makes
specifying such strings more compact.

The other main change, apart from the sequence support and target
refactoring, is an update to the 'flags' code to support a second
special tag, '_math_'. This identifies a character set based upon the
math unicode property (previously '_emoji_' was the other such special
tag).  Didn't want to support some kind of UnicodeSet-like syntax...