Change cmap reqs to support Haskell and APL in mono/math.
The character sets were discussed in issues noto-fonts#669 and
noto-fonts#751.  Some characters currently in Symbols/Symbols2
are removed, but this makes some exceptions to keep a few as
duplicates.

It is not clear whether APL support is required in Math (which is
proportional).  It will be removed if it turns out this is not
necessary.  The same might be true for Haskell.