Add tool to report missing coverage.
This is basic, it just takes the union of all the cps mentioned in a
cmap data file and compares it to defined unicode characters, minus
some pua ranges and variation selectors.  It's just to sanity check
that we didn't manage to fail to assign characters that we want to
some font.