Update noto names and data.
Instead of assuming 'Regular' is included in the full family name by
default and choosing to exclude it, assume it is excluded in the
full family name and choose to include it.  The toolchain is currently
generating full names that omit 'Regular' regardless of what the
Glyphs file says, so let's go with it.

As a result the name of one of the data fields has changed to follow
the new default behavior.

The family_name_info_p3.xml file was regenerated from a names list
generated by check_familyname_and_styles.py used with its new
data file.  family_name_info_p2.xml was edited by hand.