Use boundsPen to determine vertical extents in render.py.
This changes the types of arguments to get_glyph_cleaned_extents.

The previous code attempted to optimize computing the vertical extents
by making assumptions about what kinds of transforms would be applied
to components (translations and positive scales).  This didn't support
flips, and also didn't support rotations, which some phase3 fonts use.
BoundsPen supports the omission of single-point contours, like the old
code did, and it's more general, so rather than try to extend the old
optimizations to cover these new cases it seemed cleanest to just use
BoundsPen.

This however means the old argument types (the 'glyf' table and
_g_l_y_f.Glyph objects returned from it) needed to be converted to
_TTGlyph and TTGlyphSet, and it seems best to instead just pass them
in.

The two direct callers of this api (and the caller in render.py)
were changed to pass in the right types.

This of course breaks anyone outside of nototools who might be using
this API.