Support ExtraLight file names with weight 200.
This also changes the expectation for 'Thin' to have weight 100.

Formerly for windows GDI (which autobolds weights this low, defeating
the purpose) the minimum weight was 250 and we assigned that to 'Thin'.
The thought now is not default fonts for this use case, and instead
build custom instances for GDI when people require it.

See #334.