Generate all expected instance names from familyname_and_styles data.
Also tweak some expectations, for example the Khmer and Hebrew names
were tagged as generating only a few weights, but we actually generate
the standard range of weights for them.

Formerly we synthesized names based on the lists of masters, and made
these the 'longest' names so that the right naming properties would
be triggered.  This was especially klugy and it seemed better to
generate all of the expected style names.  To this end the sequence
of letters in which each letter named a master was changed to a small
set of tags each of which maps to a list of instance names.  The data
was modified to adjust some names accordingly.

In addition