Update unicode emoji data and tools that use it.
Only minor tweaks to unicode_data were made to handle minor format
differences.  I did not add code that processes emoji-variation-sequences
at this time, and still need to research when it might be useful.

Carried over patch to name of keycap # combining sequence, the data still
violates the expectation that '#' is globally a start-of-comment indicator.