Assume Serif fonts are not UI fonts but document fonts.
We were categorizing fonts as UI fonts based on the name or script, and
so expecting them to match a fixed ascent and descent.  If the style of
the font is 'Serif', assume it is a document font and does not need to
match the expected UI ascent and descent values.