Update emoji data parsing.
A number of emoji names are not ASCII, so this mostly tries to make that
easier by allowing emoji names to be anything (unlike Unicode
character names, which are a subset of ASCII).

The data parsing code was reading standard python 2.0 strings and not
unicode strings.  This made some operations a bit complicated, so now
data files are converted to unicode using codecs.  Of course this also
made things complicated... strings read in using codecs implement
'translate' to take fewer arguments, so the old call was failing.  To deal
with this, regex sub is used instead of translate.

The code no longer converts curly quotes to ASCII.

The sequence of operations used to collect the emoji data was reorganized.
In particular names of emoji codepoints found in the test file are
generally better than the Unicode character names (partly because they
can use apostrophe) so those are now preferred.