Expose group and order info expressed in emoji test-data file.
The emoji-test data file was previously being parsed, but only the
name data was being used (to supplement the UnicodeData information
with names for emoji that are not yet officially in that file).  This
change now parses the group and subgroup data as well, and adds apis
to make this information available, as well as information on the
overall ordering.

The test-data file did not include some sequences corresponding to
'pieces' of other sequences.  To make this useful for tools that
display these (such as the emoji html page generation tool) a 'Misc'
group was added after the existing groups and populated with this
data.  This is encoded as text inside unicode_data.py.