Change expectations for name fields for hinted phase 3 fonts.
Hinting in phase 3 is now done using ttfautohint.  It should be done
using the default ttfautohint options which appends ttfautohint version
info to the version string in the name table.

We no longer expect (in phase 3) information about hint status in the
pre-semicolon part of the version string or in the description string.