Don't require UI metrics compatibility for Serif 'deemed-UI' fonts.
Formerly lint considered fonts with UI in the name, or fonts for a short
list of scripts, to be 'UI fonts' that had to observe special metrics
constraints.  This predated Serif versions for some of these fonts.

This is now changed so fonts like Serif Hebrew aren't held to the stricter
metrics test.

The 'deemedUI' test formerly in noto_lint was moved to NotoFonts.  Font
ids still only include '-ui' if the name includes UI, and name generation
only uses 'UI' if the file name included 'UI'.  So we have two attributes,
one for UI in the name, and one for UI metrics.