is_UI_metrics include 'doc style' check, so remove unnecessary test.
The recent change to add the 'ui metrics' attribute for a font neglected
to remove the style name test code from lint that was incorporated into
that attribute.  This removes those two lines.

This also renames the ASCENT/DESCENT values, which are only applied to
UI fonts, to indicate that they're UI metrics.