Add autofix for phase 3 tool.
This is a quick oneoff tool to autofix and hint binaries for phase 3.
The idea is to take selected phase 3 fonts built by the pipeline
and promote them to the main noto-fonts hinted/unhinted directories.

Since we have existing hinted versions of some fonts, we generate
autohinted versions where necessary so both hinted and unhinted are from
the same source.

We write information into the version string so we can identify where
the font binary came from.  Currently, the binaries are built
from source and then checked in to the alpha branch of noto-fonts
without any information about the version of noto-source or the build
tools.  So the best we can do is to reflect the state of noto-fonts at
the time we took the fonts for processing.