Generate and include README files when building packages.
This fixes noto-fonts#854.

- Some tool fns were added to issue git command lines from python
and extract the required info.

- README files are generated using this info, and the corresponding
ones added to the zips when they are built.

- In the process of collecting info for the README files, we also
check to see if the repos are at a tagged commit (since we include
the tag text), and by default raise an exception if the repos are
not (or have pending changes).

Other minor changes:
- reorder the zip building processes when possible to put the
README and LICENCE files first.
- add a flag to allow continuing when repos don't pass the check
- change the name of a flag