Support a compressed syntax for representing ranges in text.
This changes parsing only.

Range parsing was simple but a bit verbose, you could either specify a
codepoint or a range of codepoints.  This is cumbersome when blocks
have lots of holes in them.

This provides a 'compressed' syntax that makes it a bit easier to
represent these kinds of ranges.  It also lifts the restriction that
range parsing always returns a set (which destroyed order-- sometimes
you want a list of characters expressed with ranges but with some
reordering).  The previous behavior remains the default, the new
behaviors have to be specifically requested.

The new syntax considers the initial string of (hex or decimal) digits
before a '-' or '/' to establish a base sequence, and subsequent ranges
of digits to replace the same number of digits in the base.  '-'
appends a range of digits between the base and the new value, '/' just
appends the new value.  The new value becomes the base for the next
'/', several of these can be strung together.  New values must be
larger than the previous value, and the 'suffix' length cannot be longer
than that of the previous 'suffix'-- this limit on functionality is an
attempt to reduce typos.